application/backend/requirements.txt
async-timeout==4.0.3
blinker==1.6.3
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.1
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.21
tldextract==5.0.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
